backend/requirements.txt
django
djangorestframework
pillow
psycopg2-binary
djangorestframework-simplejwt
python-decouple